+++ apps/web/playwright/lib/test-helpers/__tests__/routingFormHelpers.test.ts
@@ -1,4 +1,4 @@
-import prismock from "../../../../../tests/libs/__mocks__/prisma";
+import prismock from "../../../../../../tests/libs/__mocks__/prisma";
 
 import { beforeEach, describe, expect, it, vi } from "vitest";
