+++ apps/web/playwright/lib/test-helpers/teamHelpers.ts
@@ -18,12 +18,14 @@ export async function createRoundRobinTeamEventType({
   const createdEventType = await prisma.eventType.create({
     data: {
       ...eventType,
-      teamId,
       team: {
         connect: {
           id: teamId,
         },
       },
+    },
+    include: {
+      team: true,
     },
   });
