+++ packages/app-store/routing-forms/getEventTypeRedirectUrl.ts
@@ -88,6 +88,7 @@ export function getAbsoluteEventTypeRedirectUrl({
     }
   }
 
+  // We want origin to be the WEBAPP_URL as in E2E the org domain won't exist. E2E fake request from org domain using doOnOrgDomain
   const origin = process.env.NEXT_PUBLIC_IS_E2E
     ? WEBAPP_URL
     : teamSlugInRedirectUrl
